fix(parser): handle ternary after named-unary operators in statement context
The precedence chain in `parse_named_unary_statement_tail` stopped at
`parse_or_with` and never called ternary parsing. When `defined $var`
or `ref $var` appeared as a bare statement with a trailing `? then : else`,
the `?` was left over as an unexpected token.

Add `parse_ternary_with` helper that applies ternary parsing to an
existing expression node, and call it from the named-unary statement
tail between `parse_or_with` and `parse_word_or_expr`.

Fixes #1888
